use cast

+++ src/dht.hpp
@@ -254,7 +254,7 @@ namespace libcage {
                                 // reschedule
                                 timeval tval;
 
-                                tval.tv_sec  = dht::timer_interval * drand48();
+                                tval.tv_sec  = (long)((double)dht::timer_interval * drand48());
                                 tval.tv_sec += dht::timer_interval;
 
                                 tval.tv_usec = 0;
@@ -266,7 +266,7 @@ namespace libcage {
                         {
                                 timeval tval;
 
-                                tval.tv_sec  = dht::timer_interval * drand48();
+                                tval.tv_sec  = (long)((double)dht::timer_interval * drand48());
                                 tval.tv_sec += dht::timer_interval;
 
                                 tval.tv_usec = 0;
